Implement connection selection
Probability of connection is the scale of the product of
presynaptic axonal distribution and postsynaptic dendritic
distribution multiplied by appropriate length and layer priors.
Sampling connections through gsl_ran_discrete. Changed type of
length priors to double for compatibility with GSL. Corrected
scale calculation for non-overlapping distros (by convention
variance is zero when there is no probability of connection for
that class in that layer).